+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased changes
 
+- Add `Tablerone.path/1` for introspecting the local priv path of svg files.
 - `mix tablerone.download` ensures erlang dependencies are started, sets user agent.
 
 ## 0.1.0